create invoice as component and imported into main invoice route
